+++ src/prepublish.ts
@@ -1,13 +1,25 @@
 import fs from "fs";
 import path from "path";
 import cpx from "cpx";
-// import chalk from "chalk";
+import { readPackageJson } from "./internal";
+import chalk from "chalk";
+
+export interface PrepareReleaseBuildOptions {
+	filePath: string;
+	versionPlaceholder: string;
+	shouldSkip: boolean;
+}
+
+const PREPARE_RELEASE_DEFAULT_OPTIONS = Object.freeze({
+	filePath: "src/version.ts",
+	versionPlaceholder: "0.0.0-PLACEHOLDER",
+} as PrepareReleaseBuildOptions);
 
 /**
- * Prepare for prepublish, as copy files such as `README.md`, `CHANGELOG.md`, copy and transform `package.json`
+ * Build resource for publishing, copy files such as `README.md`, `CHANGELOG.md`, copy and transform `package.json`
  */
-export async function prepublish(distPath = "dist") {
-	// console.log(chalk`{blue [prepublish]} {yellow starting...}`);
+export async function buildResources(distPath = "dist") {
+	// console.log(chalk`{blue [buildResources]} {yellow starting...}`);
 
 	await writePackageTransform(distPath);
 
@@ -17,14 +29,30 @@ export async function prepublish(distPath = "dist") {
 	}
 }
 
+/** Update version etc... Generally must be invoked before building release. */
+export async function prepareReleaseBuild(options: Partial<PrepareReleaseBuildOptions> = {}) {
+	options = { ...PREPARE_RELEASE_DEFAULT_OPTIONS, ...options };
+
+	if (options.shouldSkip) {
+		console.log(chalk`{blue [prepareReleaseBuild]} {gray skipping...}`);
+		return;
+	}
+
+	await writeCodeVersion();
+}
+
 /**
  * Copy and transform package.json for publishing.
  */
 export async function writePackageTransform(distPath = "dist") {
-	const pkgKeysNormalizePaths = ["main", "umd:main", "module", "typings"];
+	const pkgKeysNormalizePaths = [
+		"main", "umd:main", "module", "typings", "es2015",
+		"esm5", "esm2015", "fesm5", "fesm2015", "metadata",
+		"esnext"
+	];
 	const pkgRemoveKeys = ["scripts", "devDependencies", "jest"];
 
-	const pkg = JSON.parse(fs.readFileSync("./package.json", "utf-8"));
+	const pkg = readPackageJson();
 
 	for (const key of pkgRemoveKeys) {
 		delete pkg[key];
@@ -44,6 +72,23 @@ export async function writePackageTransform(distPath = "dist") {
 		fs.mkdirSync(distPath);
 	}
 	await fs.promises.writeFile(path.join(distPath, "package.json"), JSON.stringify(pkg, undefined, 2));
+}
+
+/** Write package version onto code placeholder.  */
+export async function writeCodeVersion(filePath = "src/version.ts", versionPlaceholder = "0.0.0-PLACEHOLDER") {
+	if (!fs.existsSync(filePath)) {
+		return;
+	}
+
+	let fileContents = await fs.promises.readFile(filePath, { encoding: "utf-8" });
+	if (!fileContents || typeof fileContents !== "string") {
+		return;
+	}
+
+	const pkgVersion = readPackageJson().version;
+	fileContents = fileContents.replace(versionPlaceholder, pkgVersion);
+
+	await fs.promises.writeFile(filePath, fileContents);
 }
 
 function replaceAll(value: string, search: string, replacement: string): string {